Address model settings PR feedback
